+++ lib/src/pages/changelog.dart
@@ -18,6 +18,7 @@ class ChangelogPage extends StatelessWidget {
 ## Version 1.1.1
 ### **Added** 🚀
 - Added Avatars to Tasks, subtasks, comments and Kanban
+- Changed Logo design to comply with Google Play Policies
 
 ### **Fixed**  🔧
 - Kanban now refresh elements after pop out from task page.